+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - Added room activity for awake/idle state
+
 ## [15.2.2] - 2020-09-24
 ### Changed
  - Fixed room activity priorities